ai-commentary-service/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
pygame==2.5.2
